+++ CHANGELOG.md
@@ -40,9 +40,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Attribute values now validate against what is allowable per the specification
   rather than allowing anything the protobuf could encode. This may be breaking
   to some users who were relying on the incorrect behavior, such as allowing
-  dictionaries or tuples which were coerced to arrays. The one exception we have
+  dictionaries or non-homogenous lists/tuples. The one exception we have
   kept is continuing to allow atoms in place of binaries for performance.
-- Attribute values of type list must be homogenous.
+- Attribute values of type list/tuple must be homogenous.
 - Span start opts are now validated. Previously, opts underwent no validations.
 - Event and link attributes are now validated. Previously only span attributes
   were validated.